requirements.txt
fastapi
uvicorn
requests
pydantic
httpx
pyyaml
pydantic-settings
pyahocorasick
tiktoken
langdetect
textblob
aiosqlite
pytest
pytest-asyncio
